+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/notifications/handler/NotificationHandlerFactory.kt
@@ -24,10 +24,12 @@ import android.content.Context
 import android.content.Intent
 import android.os.Build
 import androidx.annotation.RequiresApi
+import androidx.core.app.NotificationCompat
 import androidx.core.graphics.drawable.IconCompat
 import io.getstream.android.push.permissions.DefaultNotificationPermissionHandler
 import io.getstream.android.push.permissions.NotificationPermissionHandler
 import io.getstream.chat.android.client.R
+import io.getstream.chat.android.client.receivers.NotificationMessageReceiver
 import io.getstream.chat.android.models.Channel
 import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.User
@@ -49,6 +51,8 @@ public object NotificationHandlerFactory {
      * Used in SDK_INT >= VERSION_CODES.O.
      * @param userIconBuilder Generates [IconCompat] to be shown on notifications.
      * @param permissionHandler Handles [android.Manifest.permission.POST_NOTIFICATIONS] permission lifecycle.
+     * @param notificationTextFormatter Lambda expression used to formats the text of the notification.
+     * @param actionsProvider Lambda expression used to provide actions for the notification.
      */
     @SuppressLint("NewApi")
     @JvmOverloads
@@ -56,57 +60,52 @@ public object NotificationHandlerFactory {
     public fun createNotificationHandler(
         context: Context,
         notificationConfig: NotificationConfig,
-        newMessageIntent: ((message: Message, channel: Channel) -> Intent)? = null,
-        notificationChannel: (() -> NotificationChannel)? = null,
+        newMessageIntent: ((message: Message, channel: Channel) -> Intent) = getDefaultNewMessageIntentFun(context),
+        notificationChannel: (() -> NotificationChannel) = getDefaultNotificationChannel(context),
         userIconBuilder: UserIconBuilder = provideDefaultUserIconBuilder(context),
         permissionHandler: NotificationPermissionHandler? = provideDefaultNotificationPermissionHandler(context),
-    ): NotificationHandler {
-        return createNotificationHandler(
+        notificationTextFormatter: ((currentUser: User?, message: Message) -> CharSequence) =
+            getDefaultNotificationTextFormatter(notificationConfig),
+        actionsProvider: (notificationId: Int, channel: Channel, message: Message) -> List<NotificationCompat.Action> =
+            getDefaultActionsProvider(context),
+    ): NotificationHandler = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+        MessagingStyleNotificationHandler(
             context = context,
             newMessageIntent = newMessageIntent,
             notificationChannel = notificationChannel,
             userIconBuilder = userIconBuilder,
             permissionHandler = permissionHandler,
-            autoTranslationEnabled = notificationConfig.autoTranslationEnabled,
+            notificationTextFormatter = notificationTextFormatter,
+            actionsProvider = actionsProvider,
+        )
+    } else {
+        ChatNotificationHandler(
+            context = context,
+            newMessageIntent = newMessageIntent,
+            notificationChannel = notificationChannel,
+            notificationTextFormatter = notificationTextFormatter,
+            actionsProvider = actionsProvider,
         )
     }
 
-    /**
-     * Method that creates a [NotificationHandler].
-     *
-     * @param context The [Context] to build the [NotificationHandler] with.
-     * @param newMessageIntent Lambda expression used to generate an [Intent] to open your app
-     * @param notificationChannel Lambda expression used to generate a [NotificationChannel].
-     * Used in SDK_INT >= VERSION_CODES.O.
-     * @param userIconBuilder Generates [IconCompat] to be shown on notifications.
-     * @param permissionHandler Handles [android.Manifest.permission.POST_NOTIFICATIONS] permission lifecycle.
-     * @param autoTranslationEnabled Enables automatic translation of push notifications.
-     */
-    @SuppressLint("NewApi")
-    @JvmOverloads
-    @JvmStatic
-    public fun createNotificationHandler(
+    private fun getDefaultActionsProvider(
         context: Context,
-        newMessageIntent: ((message: Message, channel: Channel) -> Intent)? = null,
-        notificationChannel: (() -> NotificationChannel)? = null,
-        userIconBuilder: UserIconBuilder = provideDefaultUserIconBuilder(context),
-        permissionHandler: NotificationPermissionHandler? = provideDefaultNotificationPermissionHandler(context),
-        autoTranslationEnabled: Boolean = false,
-    ): NotificationHandler {
-        val notificationChannelFun = notificationChannel ?: getDefaultNotificationChannel(context)
-        (newMessageIntent ?: getDefaultNewMessageIntentFun(context)).let { newMessageIntentFun ->
-            return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                MessagingStyleNotificationHandler(
-                    context,
-                    newMessageIntentFun,
-                    notificationChannelFun,
-                    userIconBuilder,
-                    permissionHandler,
-                    autoTranslationEnabled,
-                )
-            } else {
-                ChatNotificationHandler(context, newMessageIntentFun, notificationChannelFun, autoTranslationEnabled)
-            }
+    ): (notificationId: Int, channel: Channel, message: Message) -> List<NotificationCompat.Action> =
+        { notificationId, channel, message ->
+            listOf(
+                NotificationMessageReceiver.createReadAction(context, notificationId, channel, message),
+                NotificationMessageReceiver.createReplyAction(context, notificationId, channel),
+            )
+        }
+
+    private fun getDefaultNotificationTextFormatter(
+        notificationConfig: NotificationConfig,
+    ): (currentUser: User?, message: Message) -> CharSequence = { currentUser, message ->
+        when (notificationConfig.autoTranslationEnabled) {
+            true -> currentUser?.language?.let { userLanguage ->
+                message.getTranslation(userLanguage).ifEmpty { message.text }
+            } ?: message.text
+            else -> message.text
         }
     }