+++ stream-chat-android-ui-common/src/main/kotlin/io/getstream/chat/android/ui/common/feature/messages/composer/AudioRecordingController.kt
@@ -108,7 +108,7 @@ internal class AudioRecordingController(
                 logger.v { "[onRecorderDurationChanged] duration: $durationMs, state: $state" }
                 if (state is RecordingState.Recording) {
                     // TODO make duration Int
-                    recordingState.value = state.copy(duration = durationMs.toInt())
+                    setState(state.copy(duration = durationMs.toInt()))
                 }
             }
         }
@@ -153,7 +153,7 @@ internal class AudioRecordingController(
         val state = recordingState.value
         if (state is RecordingState.Recording) {
             logger.v { "[processWave] waveform.size($normalized): ${waveform.size}" }
-            recordingState.value = state.copy(waveform = ArrayList(waveform))
+            setState(state.copy(waveform = ArrayList(waveform)))
         }
     }
 
@@ -166,7 +166,7 @@ internal class AudioRecordingController(
         logger.i { "[startRecording] state: $state" }
         val recordingName = "audio_recording_${Date()}"
         mediaRecorder.startAudioRecording(recordingName, realPollingInterval.toLong())
-        this.recordingState.value = RecordingState.Hold(offset = offset ?: RecordingState.Hold.ZeroOffset)
+        setState(RecordingState.Hold(offset = offset ?: RecordingState.Hold.ZeroOffset))
     }
 
     public fun holdRecording(offset: Pair<Float, Float>? = null) {
@@ -180,7 +180,7 @@ internal class AudioRecordingController(
             return
         }
         logger.v { "[holdRecording] offset: Offset(${offset.first}:${offset.second})" }
-        this.recordingState.value = state.copy(offset = offset)
+        setState(state.copy(offset = offset))
     }
 
     public fun lockRecording() {
@@ -190,7 +190,7 @@ internal class AudioRecordingController(
             return
         }
         logger.i { "[lockRecording] state: $state" }
-        this.recordingState.value = RecordingState.Locked(state.durationInMs, state.waveform)
+        setState(RecordingState.Locked(state.durationInMs, state.waveform))
     }
 
     public fun cancelRecording() {
@@ -205,41 +205,42 @@ internal class AudioRecordingController(
             audioPlayer.resetAudio(state.playingId)
         }
         clearData()
-        this.recordingState.value = RecordingState.Idle
+        setState(RecordingState.Idle)
     }
 
     public fun toggleRecordingPlayback() {
         val state = this.recordingState.value
         if (state !is RecordingState.Overview) {
-            logger.w { "[toggleRecordingPlayback] rejected (state is not Locked): $state" }
+            logger.v { "[toggleRecordingPlayback] rejected (state is not Locked): $state" }
             return
         }
-        logger.i { "[toggleRecordingPlayback] state: $state" }
+        logger.i { "[toggleRecordingPlayback] state: $state, playerState: ${audioPlayer.currentState}" }
         val audioFile = state.attachment.upload ?: run {
-            logger.w { "[toggleRecordingPlayback] rejected (audioFile is null)" }
+            logger.v { "[toggleRecordingPlayback] rejected (audioFile is null)" }
             return
         }
-        if (state.isPlaying) {
-            logger.v { "[toggleRecordingPlayback] pause playback" }
-            audioPlayer.pause()
-            this.recordingState.value = state.copy(isPlaying = false)
+        if (state.hasPlayingId && state.playingId == audioPlayer.currentPlayingId) {
+            if (state.isPlaying) {
+                logger.d { "[toggleRecordingPlayback] pause playback" }
+                audioPlayer.pause()
+                setState(state.copy(isPlaying = false))
+            } else {
+                logger.d { "[toggleRecordingPlayback] resume playback" }
+                audioPlayer.resume(state.playingId)
+                setState(state.copy(isPlaying = true))
+            }
             return
         }
-        if (state.playingId != -1) {
-            logger.v { "[toggleRecordingPlayback] resume playback" }
-            // audioPlayer.play(fileToUri(audioFile), state.playingId)
-            audioPlayer.resume(state.playingId)
-            this.recordingState.value = state.copy(isPlaying = true)
-            return
-        }
-        logger.v { "[toggleRecordingPlayback] start playback" }
-        val hash = audioFile.hashCode()
-        audioPlayer.registerOnProgressStateChange(hash, ::onAudioPlayingProgress)
-        audioPlayer.registerOnAudioStateChange(hash, ::onAudioStateChanged)
-        audioPlayer.play(fileToUri(audioFile), hash)
-        this.recordingState.value = state.copy(
-            isPlaying = true,
-            playingId = hash,
+        val audioHash = audioFile.hashCode()
+        logger.d { "[toggleRecordingPlayback] start playback: $audioHash" }
+        audioPlayer.registerOnProgressStateChange(audioHash, ::onAudioPlayingProgress)
+        audioPlayer.registerOnAudioStateChange(audioHash, ::onAudioStateChanged)
+        audioPlayer.play(fileToUri(audioFile), audioHash)
+        setState(
+            state.copy(
+                isPlaying = true,
+                playingId = audioHash,
+            ),
         )
     }
 
@@ -250,25 +251,28 @@ internal class AudioRecordingController(
             return
         }
         logger.d { "[onAudioStateChanged] playbackState: $playbackState" }
-        this.recordingState.value = state.copy(
-            isPlaying = playbackState == AudioState.PLAYING,
-            playingProgress = when (playbackState) {
-                AudioState.PLAYING,
-                AudioState.PAUSE,
-                -> state.playingProgress
-                else -> 0f
-            },
+        setState(
+            state.copy(
+                isPlaying = playbackState == AudioState.PLAYING,
+                playingProgress = when (playbackState) {
+                    AudioState.PLAYING,
+                    AudioState.PAUSE,
+                    -> state.playingProgress
+                    else -> 0f
+                },
+            ),
         )
     }
 
     private fun onAudioPlayingProgress(progressState: ProgressData) {
-        // logger.d { "[onAudioPlayingProgress] progressState: $progressState" }
         val curState = this.recordingState.value
         if (curState is RecordingState.Overview) {
-            this.recordingState.value = curState.copy(
-                isPlaying = true,
-                playingProgress = progressState.progress,
-                durationInMs = progressState.duration,
+            setState(
+                curState.copy(
+                    isPlaying = true,
+                    playingProgress = progressState.progress,
+                    durationInMs = progressState.duration.takeIf { it > 0 } ?: curState.durationInMs,
+                ),
             )
         }
     }
@@ -284,7 +288,7 @@ internal class AudioRecordingController(
         if (result.isFailure) {
             logger.e { "[stopRecording] failed: ${result.errorOrNull()}" }
             clearData()
-            recordingState.value = RecordingState.Idle
+            setState(RecordingState.Idle)
             return
         }
         val adjusted = samples.downsampleMax(samplesTarget)
@@ -292,7 +296,7 @@ internal class AudioRecordingController(
         clearData()
         val recorded = result.getOrThrow()
         logger.v { "[stopRecording] recorded: $recorded" }
-        recordingState.value = RecordingState.Overview(recorded.durationInMs, normalized, recorded.attachment)
+        setState(RecordingState.Overview(recorded.durationInMs, normalized, recorded.attachment))
     }
 
     public fun seekRecordingTo(progress: Float) {
@@ -307,10 +311,9 @@ internal class AudioRecordingController(
         }
         val positionInMs = (progress * state.durationInMs).toInt()
         logger.i { "[seekRecordingTo] progress: $progress (${positionInMs}ms), state: $state" }
-        val hash = audioFile.hashCode()
-        // audioPlayer.prepare(fileToUri(audioFile), hash)
-        audioPlayer.seekTo(positionInMs, hash)
-        this.recordingState.value = state.copy(playingProgress = progress, playingId = hash)
+        val audioHash = audioFile.hashCode()
+        audioPlayer.seekTo(positionInMs, audioHash)
+        setState(state.copy(playingProgress = progress))
     }
 
     public fun pauseRecording() {
@@ -321,7 +324,7 @@ internal class AudioRecordingController(
         }
         logger.i { "[pauseRecording] state: $state" }
         audioPlayer.startSeek(state.playingId)
-        this.recordingState.value = state.copy(isPlaying = false)
+        setState(state.copy(isPlaying = false))
     }
 
     public fun completeRecording() {
@@ -335,21 +338,23 @@ internal class AudioRecordingController(
             logger.d { "[completeRecording] completing from Overview state" }
             audioPlayer.resetAudio(state.playingId)
             clearData()
-            this.recordingState.value = RecordingState.Complete(
-                state.attachment.copy(
-                    extraData = state.attachment.extraData + mapOf(
-                        EXTRA_WAVEFORM_DATA to state.waveform,
+            setState(
+                RecordingState.Complete(
+                    state.attachment.copy(
+                        extraData = state.attachment.extraData + mapOf(
+                            EXTRA_WAVEFORM_DATA to state.waveform,
+                        ),
                     ),
                 ),
             )
-            this.recordingState.value = RecordingState.Idle
+            setState(RecordingState.Idle)
             return
         }
         val result = mediaRecorder.stopRecording()
         if (result.isFailure) {
             logger.e { "[completeRecording] failed: ${result.errorOrNull()}" }
             clearData()
-            recordingState.value = RecordingState.Idle
+            setState(RecordingState.Idle)
             return
         }
         val adjusted = samples.downsampleMax(samplesTarget)
@@ -365,8 +370,8 @@ internal class AudioRecordingController(
             )
         }
         logger.d { "[completeRecording] complete from state: $state" }
-        recordingState.value = RecordingState.Complete(recorded.attachment)
-        recordingState.value = RecordingState.Idle
+        setState(RecordingState.Complete(recorded.attachment))
+        setState(RecordingState.Idle)
     }
 
     public fun onCleared() {
@@ -387,6 +392,11 @@ internal class AudioRecordingController(
         samples.clear()
         samplesBuffer.clear()
         samplesBufferLimit = 1
+        setState(RecordingState.Idle)
+    }
+
+    private fun setState(state: RecordingState) {
+        recordingState.value = state
     }
 
     private fun List<Int>.normalize(): List<Float> {